server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.12
google-generativeai==0.8.3
Pillow==10.4.0
PyPDF2==3.0.1
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2

# Development dependencies
mypy==1.11.2
black==24.8.0
types-requests==2.32.0.20241016

